refactor: split remaining oversized source and test files
Continue the file-size reduction started in #250 by splitting large
collector implementations, wizard library definitions, and route
collector tests into focused modules.

Production code:
- Java/Kotlin extended registration collectors: chain support, fluent
  route, route decorator, and virtual host helpers
- RouteProtocol enum extracted from ArmeriaRouteCollector
- ArmeriaLibraries split into category, core, and integration files

Tests:
- ArmeriaKotlinServiceRegistrationCollectorTest -> Basic, ServiceUnder,
  Variants
- ArmeriaRouteCollectorTest -> AnnotatedRoute, ServiceRegistration,
  Decorator
- ArmeriaKotlinDecoratorRouteCollectorTest -> Basic, ScopeBlock,
  PathScoped
