--- plugin-route-analysis/src/test/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaKotlinDecoratorRouteCollectorTest.kt
@@ -1,518 +1,0 @@
-package com.linecorp.intellij.plugins.armeria.explorer
-
-import com.linecorp.intellij.plugins.armeria.test.ArmeriaFixtureTestBase
-
-class ArmeriaKotlinDecoratorRouteCollectorTest : ArmeriaFixtureTestBase() {
-
-    override fun registerArmeriaStubs() {
-        registerKotlinRouteCollectorStubs()
-    }
-    fun testCollectProgrammaticDecoratorOnKotlinServiceRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.cors.CorsService
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .decorator(LoggingService::class.java)
-                    .decorator(CorsService::class.java)
-                    .service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("CORS", "Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorWithNewDecoratorFactory() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .decorator(LoggingService.newDecorator())
-                    .service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorWithBuilderNewDecoratorFactory() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .decorator(LoggingService.builder().newDecorator())
-                    .service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorInApplyBlock() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder().apply {
-                    decorator(LoggingService::class.java)
-                    service("/api", HelloService())
-                }.build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorOnKotlinBuilderVariable() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                val sb = Server.builder()
-                sb.decorator(LoggingService::class.java)
-                    .service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticPathScopedDecoratorOnKotlinServiceRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .decorator("/api/**", LoggingService::class.java)
-                    .service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectPathScopedDecoratorFiltersByRoute() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .decorator("/api/**", LoggingService::class.java)
-                    .service("/api", HelloService())
-                    .service("/other", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val apiRoute = routes.firstOrNull { it.path == "/api" }
-        val otherRoute = routes.firstOrNull { it.path == "/other" }
-        assertNotNull(apiRoute)
-        assertNotNull(otherRoute)
-        assertEquals(listOf("Logging"), apiRoute!!.decorators)
-        assertEquals(emptyList<String>(), otherRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorInAlsoBlock() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder().also {
-                    it.decorator(LoggingService::class.java)
-                    it.service("/api", HelloService())
-                }.build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorInAlsoBlockChainedAfterAlso() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder().also {
-                    it.decorator(LoggingService::class.java)
-                }.service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorFromLocalVariable() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                val logging = LoggingService.newDecorator()
-                Server.builder()
-                    .decorator(logging)
-                    .service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorInApplyBlockChainedAfterApply() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder().apply {
-                    decorator(LoggingService::class.java)
-                }.service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorInChainedApplyBlocks() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .apply { decorator(LoggingService::class.java) }
-                    .apply { service("/api", HelloService()) }
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectPathScopedDecoratorWithConstValPathPattern() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            private const val API_PATH = "/api/**"
-
-            fun main() {
-                Server.builder()
-                    .decorator(API_PATH, LoggingService::class.java)
-                    .service("/api", HelloService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorInApplyBlockStopsAfterRegistrationStatement() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.cors.CorsService
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder().apply {
-                    decorator(LoggingService::class.java)
-                    val x = service("/api", HelloService())
-                    decorator(CorsService::class.java)
-                }.build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-
-    fun testCollectProgrammaticDecoratorInApplyBlockStopsWithinRegistrationStatement() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.cors.CorsService
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder().apply {
-                    decorator(LoggingService::class.java)
-                    service("/api", HelloService()).also {
-                        decorator(CorsService::class.java)
-                    }
-                }.build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public class HelloService {
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-
-        val serviceRoute = routes.firstOrNull { it.path == "/api" && it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals(listOf("Logging"), serviceRoute!!.decorators)
-    }
-}
